refactory: replace int typedefs with cstd types

+++ include/nn/fs/fs_files.h
@@ -5,9 +5,9 @@
 namespace nn::fs {
 
 struct FileTimeStamp {
-    u64 mTime1;  // date created
-    u64 mTime2;  // sometimes is identical to above
-    u64 mTime3;  // looks like the date the file was created without exact time?
+    uint64_t mTime1;  // date created
+    uint64_t mTime2;  // sometimes is identical to above
+    uint64_t mTime3;  // looks like the date the file was created without exact time?
     bool unkBool;
 };
 
@@ -16,7 +16,7 @@ struct FileTimeStamp {
     path: Path where to create the path.
     size: Size of the file to create.
 */
-Result CreateFile(const char* path, s64 size);
+Result CreateFile(const char* path, int64_t size);
 
 /*
     Open a file.
@@ -38,7 +38,7 @@ void CloseFile(FileHandle handle);
     position:   Position within the file to be read.
     size:       How many bytes to read from file.
 */
-Result ReadFile(FileHandle handle, long position, void* buffer, ulong size);
+Result ReadFile(FileHandle handle, long position, void* buffer, size_t size);
 
 /*
     Read file at a location, with additional options.
@@ -56,7 +56,7 @@ Result ReadFile(FileHandle handle, long position, void* buffer, const ReadOption
     position:   Position within the file to be read.
     size:       How many bytes to read from file.
 */
-Result ReadFile(ulong* bytesRead, FileHandle handle, long position, void* buffer);
+Result ReadFile(size_t* bytesRead, FileHandle handle, long position, void* buffer);
 
 /*
     Read file at a location, with an output amount of bytes read, and additional options.
@@ -66,13 +66,13 @@ Result ReadFile(ulong* bytesRead, FileHandle handle, long position, void* buffer
     size:       How many bytes to read from file.
     option:     Additional options for reading, see ReadOption.
 */
-Result ReadFile(ulong* bytesRead, FileHandle handle, long position, void* buffer,
+Result ReadFile(size_t* bytesRead, FileHandle handle, long position, void* buffer,
                 const ReadOption& option);
 
-Result ReadFile(u64* outSize, FileHandle handle, s64 offset, void* buffer, u64 bufferSize,
+Result ReadFile(uint64_t* outSize, FileHandle handle, int64_t offset, void* buffer, uint64_t bufferSize,
                 const ReadOption& option);
-Result ReadFile(u64* outSize, FileHandle handle, s64 offset, void* buffer, u64 bufferSize);
-Result ReadFile(FileHandle handle, s64 offset, void* buffer, u64 bufferSize,
+Result ReadFile(uint64_t* outSize, FileHandle handle, int64_t offset, void* buffer, uint64_t bufferSize);
+Result ReadFile(FileHandle handle, int64_t offset, void* buffer, uint64_t bufferSize,
                 const ReadOption& option);
 
 /*
@@ -90,7 +90,7 @@ Result GetFileSize(long* size, nn::fs::FileHandle handle);
     size: Amount of data to write, from the pointer.
     option: Additional options for writing, like flushing.
 */
-Result WriteFile(FileHandle handle, s64 position, void const* buffer, u64 size,
+Result WriteFile(FileHandle handle, int64_t position, void const* buffer, uint64_t size,
                  WriteOption const& option);
 
 /*
@@ -99,13 +99,13 @@ Result WriteFile(FileHandle handle, s64 position, void const* buffer, u64 size,
 */
 Result FlushFile(FileHandle handle);
 
-// Result GetSaveDataTimeStamp(nn::time::PosixTime *,ulong);
-// Result GetSaveDataTimeStamp(nn::time::PosixTime*, nn::fs::SaveDataSpaceId, ulong);
+// Result GetSaveDataTimeStamp(nn::time::PosixTime *,size_t);
+// Result GetSaveDataTimeStamp(nn::time::PosixTime*, nn::fs::SaveDataSpaceId, size_t);
 Result GetFileTimeStampForDebug(nn::fs::FileTimeStamp*, const char*);
 
 Result DeleteFile(const char* path);
 
-Result SetFileSize(FileHandle fileHandle, s64 filesize);
+Result SetFileSize(FileHandle fileHandle, int64_t filesize);
 Result RenameFile(const char* filepath, const char* newPath);
 
 }  // namespace nn::fs
